examen microservicio stock
