+++ code/orden-ms/src/main/java/com/nttdata/orden_ms/model/dto/UserSaveRequestDto.java
@@ -1,0 +1,13 @@
+package com.nttdata.orden_ms.model.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserSaveRequestDto {
+
+    private String name;
+
+    private String email;
+}
