+++ code/orden-ms/src/main/java/com/nttdata/orden_ms/model/dto/UserSaveResponseDto.java
@@ -1,0 +1,19 @@
+package com.nttdata.orden_ms.model.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Getter
+@Setter
+public class UserSaveResponseDto {
+
+    private Long id;
+
+    private String name;
+
+    private String email;
+
+    private Date registrationDate;
+}
