+++ code/orden-ms/src/main/java/com/nttdata/orden_ms/model/dto/UserUpdateRequestDto.java
@@ -1,0 +1,16 @@
+package com.nttdata.orden_ms.model.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Getter
+@Setter
+public class UserUpdateRequestDto {
+
+    private String name;
+    private String email;
+    private Boolean active;
+    private Date registrationDate;
+}
