+++ code/orden-ms/src/main/java/com/nttdata/orden_ms/model/entity/OrderDetail.java
@@ -19,8 +19,7 @@ public class OrderDetail {
     @JoinColumn(name = "order_id", nullable = false)
     private Order order;
 
-    @ManyToOne
-    @JoinColumn(name = "product_id", nullable = false)
+    @Column(nullable = false)
     private Long productId;
 
     @Column(nullable = false)
