+++ code/orden-ms/src/main/java/com/nttdata/orden_ms/model/entity/User.java
@@ -1,0 +1,38 @@
+package com.nttdata.orden_ms.model.entity;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.util.Date;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+@Entity
+@Table(name = "users")
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false, unique = true)
+    private String email;
+
+    @Column
+    private Boolean active;
+
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    @Column(nullable = false, updatable = false)
+    @CreationTimestamp
+    private Date registrationDate;
+}
