+++ code/orden-ms/src/main/java/com/nttdata/orden_ms/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.nttdata.orden_ms.repository;
+
+import com.nttdata.orden_ms.model.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
