+++ code/product-ms/src/main/java/com/nttdata/product_ms/controller/CategoryController.java
@@ -22,7 +22,6 @@ public class CategoryController {
     @Autowired
     private CategoryService categoryService;
 
-
     @GetMapping
     public List<CategoryDto>  getAllCategories() {
         return INSTANCE.map(categoryService.listAll());
