+++ code/stock-ms/src/main/java/com/nttdata/stock_ms/StockMsApplication.java
@@ -1,0 +1,13 @@
+package com.nttdata.stock_ms;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StockMsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(StockMsApplication.class, args);
+	}
+
+}
