+++ code/stock-ms/src/main/java/com/nttdata/stock_ms/exception/BadRequestException.java
@@ -1,0 +1,7 @@
+package com.nttdata.stock_ms.exception;
+
+public class BadRequestException extends RuntimeException{
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
