+++ code/stock-ms/src/main/java/com/nttdata/stock_ms/exception/InternalServerErrorException.java
@@ -1,0 +1,7 @@
+package com.nttdata.stock_ms.exception;
+public class InternalServerErrorException extends RuntimeException {
+    public InternalServerErrorException(String message) {
+        super(message);
+    }
+}
+
