+++ code/stock-ms/src/main/java/com/nttdata/stock_ms/exception/NotFoundException.java
@@ -1,0 +1,16 @@
+package com.nttdata.stock_ms.exception;
+
+public class NotFoundException extends  RuntimeException{
+
+    private final String code;
+
+    public NotFoundException(String message, String code) {
+        super(message);
+        this.code = code;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+}
