+++ code/stock-ms/src/main/java/com/nttdata/stock_ms/model/dto/ErrorDetailDto.java
@@ -1,0 +1,10 @@
+package com.nttdata.stock_ms.model.dto;
+
+import lombok.Data;
+import java.time.LocalDateTime;
+
+@Data
+public class ErrorDetailDto {
+    private String message;
+    private LocalDateTime dateTime;
+}
