+++ code/stock-ms/src/main/java/com/nttdata/stock_ms/model/dto/FindByProductIdDto.java
@@ -1,0 +1,9 @@
+package com.nttdata.stock_ms.model.dto;
+
+import lombok.Data;
+
+@Data
+public class FindByProductIdDto {
+    private Integer productId;
+    private Integer total;
+}
