+++ code/stock-ms/src/main/java/com/nttdata/stock_ms/model/dto/SaveStockRequestDto.java
@@ -1,0 +1,12 @@
+package com.nttdata.stock_ms.model.dto;
+
+import lombok.Data;
+
+@Data
+public class SaveStockRequestDto {
+
+    private Integer productId;
+    private Integer wareHouseId;
+    private Integer quantity;
+
+}
