+++ code/stock-ms/src/main/resources/application.properties
@@ -1,0 +1,20 @@
+spring.application.name=stock-ms
+
+# Puerto de la aplicación
+server.port=8095
+
+# Configuración de H2
+spring.datasource.url=jdbc:h2:file:./data/stock-db;DB_CLOSE_ON_EXIT=FALSE;AUTO_RECONNECT=TRUE
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+# Dialecto de Hibernate para H2
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+
+# Crear y actualizar las tablas automáticamente
+spring.jpa.hibernate.ddl-auto=update
+
+# Consola H2 habilitada
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
