+++ code/stock-ms/src/test/java/com/nttdata/stock_ms/StockMsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.nttdata.stock_ms;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class StockMsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
