--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/UserAndCategoryMsApplication.java
@@ -1,13 +1,0 @@
-package com.nttdata.user_and_category_ms;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserAndCategoryMsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserAndCategoryMsApplication.class, args);
-	}
-
-}
