--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/exception/BadRequestException.java
@@ -1,7 +1,0 @@
-package com.nttdata.user_and_category_ms.exception;
-
-public class BadRequestException extends RuntimeException{
-    public BadRequestException(String message) {
-        super(message);
-    }
-}
