--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/exception/InternalServerErrorException.java
@@ -1,8 +1,0 @@
-package com.nttdata.user_and_category_ms.exception;
-
-public class InternalServerErrorException extends RuntimeException {
-    public InternalServerErrorException(String message) {
-        super(message);
-    }
-}
-
