--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/exception/NotFoundException.java
@@ -1,16 +1,0 @@
-package com.nttdata.user_and_category_ms.exception;
-
-public class NotFoundException extends  RuntimeException{
-
-    private final String code;
-
-    public NotFoundException(String message, String code) {
-        super(message);
-        this.code = code;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-}
