--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/mapper/UserMapper.java
@@ -1,34 +1,0 @@
-package com.nttdata.user_and_category_ms.mapper;
-
-import com.nttdata.dockerized.postgresql.model.dto.UserDto;
-import com.nttdata.dockerized.postgresql.model.dto.UserSaveRequestDto;
-import com.nttdata.dockerized.postgresql.model.dto.UserSaveResponseDto;
-import com.nttdata.dockerized.postgresql.model.dto.UserUpdateRequestDto;
-import com.nttdata.dockerized.postgresql.model.entity.User;
-import org.mapstruct.AfterMapping;
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingTarget;
-import org.mapstruct.factory.Mappers;
-
-import java.util.List;
-
-@Mapper
-public interface UserMapper {
-
-    UserMapper INSTANCE = Mappers.getMapper(UserMapper.class);
-
-    public UserDto map(User user);
-
-    public List<UserDto> map(List<User> users);
-
-    public User toEntity(UserSaveRequestDto userSaveRequestDto);
-
-    public User toEntity(UserUpdateRequestDto userUpdateRequestDto);
-
-    public UserSaveResponseDto toUserSaveResponseDto(User user);
-
-    @AfterMapping
-    default void setRemainingValues(User user, @MappingTarget UserDto userDto) {
-        userDto.setStatus(Boolean.TRUE.equals(user.getActive()) ? "Active" : "Inactive");
-    }
-}
