--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/model/dto/UserSaveRequestDto.java
@@ -1,13 +1,0 @@
-package com.nttdata.user_and_category_ms.model.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserSaveRequestDto {
-
-    private String name;
-
-    private String email;
-}
