--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/model/dto/UserSaveResponseDto.java
@@ -1,19 +1,0 @@
-package com.nttdata.user_and_category_ms.model.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.Date;
-
-@Getter
-@Setter
-public class UserSaveResponseDto {
-
-    private Long id;
-
-    private String name;
-
-    private String email;
-
-    private Date registrationDate;
-}
