--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/model/entity/User.java
@@ -1,38 +1,0 @@
-package com.nttdata.user_and_category_ms.model.entity;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.util.Date;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Setter
-@Getter
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Column
-    private String name;
-
-    @Column
-    private String email;
-
-    @Column
-    private Boolean active;
-
-    @JsonFormat(pattern = "yyyy-MM-dd")
-    @Column
-    @CreationTimestamp
-    private Date registrationDate;
-}
