--- code/user-and-category-ms/src/main/java/com/nttdata/user_and_category_ms/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.nttdata.user_and_category_ms.repository;
-
-import com.nttdata.dockerized.postgresql.model.entity.User;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-}
