--- code/user-and-category-ms/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.application.name=user-and-category-ms
