--- code/user-and-category-ms/src/test/java/com/nttdata/user_and_category_ms/UserAndCategoryMsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.nttdata.user_and_category_ms;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserAndCategoryMsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
